requirements.txt
Pillow
numpy
python-docx
tqdm
pypandoc